+++ packages/utils/CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Minor Changes
 
 - Append data in a `FormData` instance when using an axios instance created by `createAxiosInstance` setting `file` and `useFormData` options to `true`. This will be the default behavior when using `axiosForFiles`.
-- Create `filterDirtyValues` and `getInitialValues` util functions.
+- Create `filterDirtyValues`, `getInitialValues` and `toBase64` util functions.
 - Add default parameters to the `refreshAccessToken` function.
 
 ## 2.0.0